Refactored ftrace_helper.h
